jwt implementado

+++ src/services/pedidoServices.go
@@ -2,6 +2,7 @@ package services
 
 import (
 	"ApiRestaurant/src/entity"
+	"ApiRestaurant/src/utils"
 	"database/sql"
 	"encoding/json"
 	"errors"
@@ -38,17 +39,11 @@ func (ps *PedidoServices) GetAllPedidos() (*[]entity.Pedido, error) {
 			return nil, err
 		}
 
-		hora, err := time.Parse("15:04:05", horaStr) // Formato para hora
+		err := ps.parseTime(&pedido, horaStr, fechaStr)
 		if err != nil {
 			return nil, err
 		}
-		pedido.Hora = hora
 
-		fecha, err := time.Parse("2006-01-02", fechaStr) // Formato para fecha
-		if err != nil {
-			return nil, err
-		}
-		pedido.Fecha = fecha
 		var hamburguesas []entity.Hamburguesa
 		if err := json.Unmarshal(hamburgesaJSON, &hamburguesas); err != nil {
 			return nil, err
@@ -62,18 +57,24 @@ func (ps *PedidoServices) GetAllPedidos() (*[]entity.Pedido, error) {
 	return &pedidos, nil
 }
 
-func (ps *PedidoServices) GetById(id int) (*entity.Pedido, error) {
+func (ps *PedidoServices) GetById(id int, userId int) (*entity.Pedido, error) {
 	if ps.DB == nil {
 		return nil, errors.New("la conexion a la base de datos es nil")
 	}
-
 	var pedido entity.Pedido
 	var hamburgesaJSON []byte
 	var horaStr string
 	var fechaStr string
-
-	err := ps.DB.QueryRow("SELECT * FROM pedidos WHERE pedido_id = ?", id).Scan(&pedido.Id, &pedido.UserId, &pedido.Direccion, &pedido.Price, &pedido.State, &hamburgesaJSON, &horaStr, &fechaStr)
-
+	rolId := utils.GetUserRoles(userId, ps.DB)
+	var err error
+	if rolId == 8 {
+		err = ps.DB.QueryRow("SELECT * FROM pedidos WHERE pedido_id = ?", id).Scan(&pedido.Id, &pedido.UserId, &pedido.Direccion, &pedido.Price, &pedido.State, &hamburgesaJSON, &horaStr, &fechaStr)
+	} else {
+		err = ps.DB.QueryRow("SELECT * FROM pedidos WHERE pedido_id = ? AND user_id = ?", id, userId).Scan(&pedido.Id, &pedido.UserId, &pedido.Direccion, &pedido.Price, &pedido.State, &hamburgesaJSON, &horaStr, &fechaStr)
+		if pedido.UserId != userId {
+			return nil, errors.New("El pedido no pertenece a este usuario")
+		}
+	}
 	if err != nil {
 		if err == sql.ErrNoRows {
 			return nil, errors.New("Pedido not found")
@@ -81,17 +82,10 @@ func (ps *PedidoServices) GetById(id int) (*entity.Pedido, error) {
 		return nil, err
 	}
 
-	hora, err := time.Parse("15:04:05", horaStr) // Formato para hora
+	err = ps.parseTime(&pedido, horaStr, fechaStr)
 	if err != nil {
 		return nil, err
 	}
-	pedido.Hora = hora
-
-	fecha, err := time.Parse("2006-01-02", fechaStr) // Formato para fecha
-	if err != nil {
-		return nil, err
-	}
-	pedido.Fecha = fecha
 
 	var hamburguesas []entity.Hamburguesa
 	if err := json.Unmarshal(hamburgesaJSON, &hamburguesas); err != nil {
@@ -124,17 +118,10 @@ func (ps *PedidoServices) GetByUserId(user_id int) (*[]entity.Pedido, error) {
 			return nil, err
 		}
 
-		hora, err := time.Parse("15:04:05", horaStr)
+		err := ps.parseTime(&pedido, horaStr, fechaStr)
 		if err != nil {
 			return nil, err
 		}
-		pedido.Hora = hora
-
-		fecha, err := time.Parse("2006-01-02", fechaStr)
-		if err != nil {
-			return nil, err
-		}
-		pedido.Fecha = fecha
 
 		var hamburguesas []entity.Hamburguesa
 		if err := json.Unmarshal(hamburgesaJSON, &hamburguesas); err != nil {
@@ -165,22 +152,14 @@ func (ps *PedidoServices) GetByFecha(fecha string) (*[]entity.Pedido, error) {
 		var hamburgesaJSON []byte
 		var horaStr string
 		var fechaStr string
-
 		if err := rows.Scan(&pedido.Id, &pedido.UserId, &pedido.Direccion, &pedido.Price, &pedido.State, &hamburgesaJSON, &horaStr, &fechaStr); err != nil {
 			return nil, err
 		}
 
-		hora, err := time.Parse("15:04:05", horaStr)
+		err := ps.parseTime(&pedido, horaStr, fechaStr)
 		if err != nil {
 			return nil, err
 		}
-		pedido.Hora = hora
-
-		fecha, err := time.Parse("2006-01-02", fechaStr)
-		if err != nil {
-			return nil, err
-		}
-		pedido.Fecha = fecha
 
 		var hamburguesas []entity.Hamburguesa
 		if err := json.Unmarshal(hamburgesaJSON, &hamburguesas); err != nil {
@@ -219,11 +198,19 @@ func (ps *PedidoServices) CreatePedido(pedido *entity.Pedido) error {
 	return nil
 }
 
-func (ps *PedidoServices) DeletePedido(id int) error {
+func (ps *PedidoServices) DeletePedido(id int, userId int) error {
 	if ps.DB == nil {
 		return errors.New("la conexion a la base de datos es nil")
 	}
-	result, err := ps.DB.Exec("DELETE FROM pedidos WHERE pedido_id = ?", id)
+
+	var err error
+	var result sql.Result
+	roleId := utils.GetUserRoles(userId, ps.DB)
+	if roleId == 8 {
+		result, err = ps.DB.Exec("DELETE FROM pedidos WHERE pedido_id = ?", id)
+	} else {
+		result, err = ps.DB.Exec("DELETE FROM pedidos WHERE pedido_id = ? and user_id = ?", id, userId)
+	}
 	if err != nil {
 		return err
 	}
@@ -232,19 +219,19 @@ func (ps *PedidoServices) DeletePedido(id int) error {
 		return err
 	}
 	if rowsAffected == 0 {
-		return fmt.Errorf("ningún pedido encontrado con el id %d", id)
+		return fmt.Errorf("El pedido no pudo ser borrado")
 	}
 
 	return nil
 }
 
-func (ps *PedidoServices) UpdatePedido(id int, pedido entity.Pedido) error {
+func (ps *PedidoServices) UpdatePedido(id int, pedido entity.Pedido, userId int) error {
 	if ps.DB == nil {
 		return errors.New("la conexion a la base de datos es nil")
 	}
-	existingPedido, _ := ps.GetById(id)
+	existingPedido, _ := ps.GetById(id, userId)
 	if existingPedido == nil {
-		err := fmt.Errorf("pedido no encontrado con el id %d", id)
+		err := fmt.Errorf("Pedido no encontrado con el id %d", id)
 		return err
 	}
 	hamburguesasJson, err := json.Marshal(pedido.Hamburguesas)
@@ -255,5 +242,21 @@ func (ps *PedidoServices) UpdatePedido(id int, pedido entity.Pedido) error {
 	if err1 != nil {
 		return err1
 	}
+	return nil
+}
+
+func (ps *PedidoServices) parseTime(pedido *entity.Pedido, horaStr string, fechaStr string) error {
+
+	hora, err := time.Parse("15:04:05", horaStr)
+	if err != nil {
+		return err
+	}
+	pedido.Hora = hora
+
+	fecha, err := time.Parse("2006-01-02", fechaStr)
+	if err != nil {
+		return err
+	}
+	pedido.Fecha = fecha
 	return nil
 }
